+++ src/Exceptions/InvalidFormatArgumentException.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Myerscode\Utilities\Strings\Exceptions;
-
-use InvalidArgumentException;
-
-class InvalidFormatArgumentException extends InvalidArgumentException
-{
-}